requirements.txt
annotated-types==0.7.0
anyio==4.6.0
click==8.1.7
colorama==0.4.6
fastapi==0.115.0
h11==0.14.0
idna==3.10
MarkupSafe==3.0.1
pydantic==2.9.2
pydantic_core==2.23.4
python-multipart==0.0.12
sniffio==1.3.1
starlette==0.38.6
typing_extensions==4.12.2
uvicorn==0.31.0
Werkzeug==3.0.4
certifi==2024.8.30
charset-normalizer==3.3.2
colorama==0.4.6
filelock==3.13.1
fsspec==2024.2.0
idna==3.10
Jinja2==3.1.3
llvmlite==0.43.0
MarkupSafe==3.0.1
mpmath==1.3.0
networkx==3.2.1
numba==0.60.0
numpy==1.26.3
openai-whisper==20240930
regex==2024.9.11
requests==2.32.3
setuptools==70.0.0
sympy==1.12
tiktoken==0.8.0
torch==2.4.1+cu121
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.3
websockets==13.1